Move time-related things into a separate module
