+++ tracing-fmt/src/default.rs
@@ -1,12 +1,9 @@
 use span;
+use time::{self, FormatTime, SystemTime};
 use FormatEvent;
-
-#[cfg(feature = "chrono")]
-use chrono;
 
 use std::fmt::{self, Write};
 use std::marker::PhantomData;
-use std::time::Instant;
 use tracing_core::{
     field::{self, Field},
     Event, Level,
@@ -22,84 +19,6 @@ pub struct Compact;
 /// Marker for `Format` and `Builder` that indicates that the verbose log format should be used.
 #[derive(Default, Debug, Copy, Clone, Eq, PartialEq)]
 pub struct Full;
-
-/// A type that can measure and format the current time.
-///
-/// This trait is used by `Format` to include a timestamp with each `Event` when it is logged.
-///
-/// Notable default implementations of this trait are `SystemTime` and `()`. The former prints the
-/// current time as reported by `std::time::SystemTime`, and the latter does not print the current
-/// time at all. `FormatTime` is also automatically implemented for any function pointer with the
-/// appropriate signature.
-pub trait FormatTime {
-    /// Measure and write out the current time.
-    ///
-    /// When `format_time` is called, implementors should get the current time using their desired
-    /// mechanism, and write it out to the given `fmt::Write`. Implementors must insert a trailing
-    /// space themselves if they wish to separate the time from subsequent log message text.
-    fn format_time(&self, w: &mut fmt::Write) -> fmt::Result;
-}
-
-impl FormatTime for () {
-    fn format_time(&self, _: &mut fmt::Write) -> fmt::Result {
-        Ok(())
-    }
-}
-
-impl FormatTime for fn(&mut fmt::Write) -> fmt::Result {
-    fn format_time(&self, w: &mut fmt::Write) -> fmt::Result {
-        (*self)(w)
-    }
-}
-
-/// Retrieve and print the current wall-clock time.
-///
-/// If the `chrono` feature is enabled, the current time is printed in a human-readable format like
-/// "Jun 25 14:27:12.955". Otherwise the `Debug` implementation of `std::time::SystemTime` is used.
-#[derive(Debug, Clone, Copy, Eq, PartialEq, Default)]
-pub struct SystemTime;
-
-/// Retrieve and print the relative elapsed wall-clock time since an epoch.
-///
-/// The `Default` implementation for `Uptime` makes the epoch the current time.
-#[derive(Debug, Clone, Copy, Eq, PartialEq)]
-pub struct Uptime {
-    epoch: Instant,
-}
-
-impl Default for Uptime {
-    fn default() -> Self {
-        Uptime {
-            epoch: Instant::now(),
-        }
-    }
-}
-
-impl From<Instant> for Uptime {
-    fn from(epoch: Instant) -> Self {
-        Uptime { epoch }
-    }
-}
-
-#[cfg(feature = "chrono")]
-impl FormatTime for SystemTime {
-    fn format_time(&self, w: &mut fmt::Write) -> fmt::Result {
-        write!(w, "{} ", chrono::Local::now().format("%b %d %H:%M:%S%.3f"))
-    }
-}
-#[cfg(not(feature = "chrono"))]
-impl FormatTime for SystemTime {
-    fn format_time(&self, w: &mut fmt::Write) -> fmt::Result {
-        write!(w, "{:?} ", std::time::SystemTime::now())
-    }
-}
-
-impl FormatTime for Uptime {
-    fn format_time(&self, w: &mut fmt::Write) -> fmt::Result {
-        let e = self.epoch.elapsed();
-        write!(w, "{}.{:09} ", e.as_secs(), e.subsec_nanos())
-    }
-}
 
 /// Builder for a `Format` formatter.
 #[derive(Debug, Clone)]
@@ -169,23 +88,6 @@ impl<T: Default> Default for Format<Full, T> {
     }
 }
 
-impl<F, T> Format<F, T>
-where
-    T: FormatTime,
-{
-    #[inline(always)]
-    fn time(&self, writer: &mut fmt::Write) -> fmt::Result {
-        #[cfg(feature = "ansi")]
-        let style = Style::new().dimmed();
-        #[cfg(feature = "ansi")]
-        write!(writer, "{}", style.prefix())?;
-        self.timer.format_time(writer)?;
-        #[cfg(feature = "ansi")]
-        write!(writer, "{}", style.suffix())?;
-        Ok(())
-    }
-}
-
 impl<N, T> FormatEvent<N> for Format<Full, T>
 where
     N: for<'a> ::NewVisitor<'a>,
@@ -198,7 +100,7 @@ where
         event: &Event,
     ) -> fmt::Result {
         let meta = event.metadata();
-        self.time(writer)?;
+        time::write(&self.timer, writer)?;
         write!(
             writer,
             "{} {}{}: ",
@@ -226,7 +128,7 @@ where
         event: &Event,
     ) -> fmt::Result {
         let meta = event.metadata();
-        self.time(writer)?;
+        time::write(&self.timer, writer)?;
         write!(
             writer,
             "{} {}{}: ",
